src/backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
streamlit
passlib
